updated vesion fixed bugs

+++ MatrixSolution.cpp
@@ -30,5 +30,5 @@ std::string MatrixSolution::recursion_path(State<double> *state) {
   else {
     direction = UP;
   }
-  return recursion_path((*state).get_parent()) + direction + " (" + std::to_string((*state).get_cost()) + ") ,";
+  return recursion_path((*state).get_parent()) + direction + " (" + std::to_string((int)(*state).get_cost()) + ") ,";
 }
